make startActivity() handle both setPackage() and setClassName()
An Intent can be targeted using both setPackage() and setClassName().
setPackage() narrows down the possible recipients to a given packageName,
while setClassName() set the specific Activity in an app to send to. The
trust check is based on packageName since that is what the APK signature is
tied to.  So if getPackage() turns up nothing, then try getComponent(),
which fetches the CompenentName which is a representation of the thing that
was set using Intent.setClassName()